+++ tools/src/coverage/CoverageCalculator.ts
@@ -8,7 +8,7 @@
 */
 
 import { type OpenAPIV3 } from 'openapi-types'
-import { HTTP_METHODS, read_yaml, write_json } from '../../helpers'
+import { HTTP_METHODS, read_yaml, write_json } from '../helpers'
 
 export default class CoverageCalculator {
   private readonly _cluster_spec: OpenAPIV3.Document